requirements.txt
# LLM
google-generativeai>=0.8.0

# RAG
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
langchain-text-splitters>=0.2.0

# Motor
# (sin dependencias externas - Python puro)

# Frontend
streamlit>=1.38.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
